Various fixes in backend access.

+++ tchan2/backend/backend.go
@@ -14,11 +14,11 @@ type DB interface {
 	// PopulateBoard fetches a board by name.
 	PopulateBoard(boardName string, b *tchan2.BoardOverview, ok *bool) error
 	// PopulateThread fetches the thread with the specified post in it.
-	PopulateThread(boardName string, postID int, thr *tchan2.Thread, ok *bool) error
+	PopulateThread(boardName string, postID int64, thr *tchan2.Thread, ok *bool) error
 	// CreateThread adds a new thread to a board, setting the OP's post ID.
 	CreateThread(boardName string, topic string, op *tchan2.Post) error
 	// AddPostToThread adds a reply to a thread, setting the post's ID in the process.
-	AddReply(boardName string, postID int, post *tchan2.Post, ok *bool) error
+	AddReply(boardName string, postID int64, post *tchan2.Post, ok *bool) error
 }
 
 // New creates a new backend which has yet to be initialized.
